requirements.txt
asyncvnc>=1.3.0,<2
fastapi>=0.128.0,<1
uvicorn>=0.40.0,<1
numpy>=2.4.2,<3
Pillow>=12.1.1,<13
